Move blob sender metric to the separate module
